Saved Valid Post Data to the Database
I have managed to save valid post data to the database and am now working
on how to handle errors from the databse. For example if a user submits
data using an email that is already in the database it should not raise an
error "UNIQUE constraint failed: mainapp_user.email". Instead it should
display a form with the errors.