Saving Post Data to the Database
I have managed to save registered users to the database and returning a
form with an error when a user tries to register with an email adress that
already exists in the databse. I now need to use password hashing to
encrypt the passwords before storing them in the database.