Saving Question to the Database
I was able to save questions to the database. I had many errors previously but i could not solve any error untill i realized that the name of a survey recieved from the front-end the second time did not have any spaces, while the names in the database had spaces. I changed the contents of the databse manually. I now need to solve the survey naming issue before it occurs.